# Add module level SkipLocalsInit