--- a/Fluid/Properties/AssemblyInfo.cs
+++ b/Fluid/Properties/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[module: System.Runtime.CompilerServices.SkipLocalsInit]